fix: Its being added the new local provider structure
